Updated parameter types and names to match Arnold

+++ props/options.py
@@ -46,7 +46,7 @@ class ArnoldOptions(PropertyGroup):
         min=0, 
         default=0
         )
-    sample_clamp: IntProperty(
+    sample_clamp: FloatProperty(
         name="Sample Clamp",
         description="If enabled, this control will clamp pixel samples to this specified maximum value. This can make it easier to anti-alias certain high-dynamic-range effects such as bright motion-blurred specular streaks (at the cost of reduced contrast)",
         min=0,
@@ -56,7 +56,7 @@ class ArnoldOptions(PropertyGroup):
         name="Clamp AOVs",
         description="With this control enabled the pixel samples of the AOVs will also be clamped. AOV clamping will affect every RGB and RGBA (except the A component) AOV. Currently, there is no way to tell Arnold which AOV's to clamp and which ones to preserve"
         )
-    indirect_sample_clamp: IntProperty(
+    indirect_sample_clamp: FloatProperty(
         name="Indirect Sample Clamp",
         description="The threshold to clamp away fireflies from indirect light samples and reduce noise. This works similarly to AA_sample_clamp but preserves specular highlights from direct lighting. Lower values result in more aggressive noise reduction, possibly at the expense of dynamic range",
         min=0,
